+++ src/app/app-wrapper.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
+import { node } from 'prop-types';
 
 import ApiWrapper from 'api/api-wrapper';
+import BrowserWrapper from 'components/browser/browser-wrapper';
 import CmsWrapper from 'components/cms-widget/cms-wrapper';
 import CompareWrapper from 'components/compare-widget/compare-wrapper';
 import FlagWrapper from 'api/flag-wrapper';
@@ -13,7 +15,9 @@ function AppWrapper({ children }) {
         <FlagWrapper>
           <CompareWrapper>
             <CmsWrapper>
-              {children}
+              <BrowserWrapper>
+                {children}
+              </BrowserWrapper>
             </CmsWrapper>
           </CompareWrapper>
         </FlagWrapper>
@@ -23,3 +27,7 @@ function AppWrapper({ children }) {
 }
 
 export default AppWrapper;
+
+AppWrapper.propTypes = {
+  children: node.isRequired,
+};