+++ src/app/pages/administration/search/search-view.jsx
@@ -1,7 +1,5 @@
 import React, { useEffect, useState } from 'react';
 import {
-  Button,
-  ButtonGroup,
   Paper,
   Table,
   TableBody,
@@ -12,12 +10,11 @@ import {
   makeStyles,
 } from '@material-ui/core';
 import { shape, string } from 'prop-types';
-import GetAppIcon from '@material-ui/icons/GetApp';
-import { ExportToCsv } from 'export-to-csv';
 
 import { useFetchCollection } from 'api/collections';
-import ChplCertificationStatusLegend from 'components/certification-status/certification-status';
 import ChplActionButton from 'components/action-widget/action-button';
+import ChplCertificationStatusLegend from 'components/certification-status/certification-status';
+import ChplDownloadListings from 'components/download-listings/download-listings';
 import {
   ChplLink,
   ChplPagination,
@@ -26,6 +23,7 @@ import { ChplSortableHeaders } from 'components/util/sortable-headers';
 import {
   ChplFilterChips,
   ChplFilterPanel,
+  ChplFilterQuickFilters,
   ChplFilterSearchTerm,
   useFilterContext,
 } from 'components/filter';
@@ -35,21 +33,7 @@ import { getDisplayDateFormat } from 'services/date-util';
 import { useSessionStorage as useStorage } from 'services/storage.service';
 import { palette, theme } from 'themes';
 
-const csvOptions = {
-  filename: 'listings',
-  showLabels: true,
-  headers: [
-    { headerName: 'CHPL ID', objectKey: 'chplProductNumber' },
-    { headerName: 'Certification Edition', objectKey: 'fullEdition' },
-    { headerName: 'Developer', objectKey: 'developerName' },
-    { headerName: 'Product', objectKey: 'productName' },
-    { headerName: 'Version', objectKey: 'versionName' },
-    { headerName: 'Certification Date', objectKey: 'certificationDate' },
-    { headerName: 'Certification Status', objectKey: 'certificationStatusName' },
-  ],
-};
-
-/* eslint-disable object-curly-newline */
+/* eslint object-curly-newline: ["error", { "minProperties": 5, "consistent": true }] */
 const headers = [
   { property: 'chpl_id', text: 'CHPL ID', sortable: true },
   { text: 'Certification Edition' },
@@ -60,12 +44,8 @@ const headers = [
   { text: 'Status', extra: <ChplCertificationStatusLegend /> },
   { text: 'Actions', invisible: true },
 ];
-/* eslint-enable object-curly-newline */
 
 const useStyles = makeStyles({
-  iconSpacing: {
-    marginLeft: '4px',
-  },
   linkWrap: {
     overflowWrap: 'anywhere',
   },
@@ -91,7 +71,7 @@ const useStyles = makeStyles({
     gap: '16px',
     alignItems: 'center',
     [theme.breakpoints.up('md')]: {
-      gridTemplateColumns: 'auto 10fr auto',
+      gridTemplateColumns: 'auto 10fr auto auto',
     },
   },
   stickyColumn: {
@@ -137,7 +117,7 @@ function ChplSearchView(props) {
   const storageKey = 'storageKey-searchView';
   const $analytics = getAngularService('$analytics');
   const { analytics } = props;
-  const csvExporter = new ExportToCsv(csvOptions);
+  const [directReviewsAvailable, setDirectReviewsAvailable] = useState(true);
   const [listings, setListings] = useState([]);
   const [orderBy, setOrderBy] = useStorage(`${storageKey}-orderBy`, 'developer');
   const [pageNumber, setPageNumber] = useStorage(`${storageKey}-pageNumber`, 0);
@@ -161,16 +141,12 @@ function ChplSearchView(props) {
       setListings([]);
       return;
     }
+    setDirectReviewsAvailable(data?.directReviewsAvailable);
     setListings(data.results.map((listing) => ({
       ...listing,
-      fullEdition: `${listing.edition.name}${listing.curesUpdate ? ' Cures Update' : ''}`,
-      developerName: listing.developer.name,
-      productName: listing.product.name,
-      versionName: listing.version.name,
-      certificationStatusName: listing.certificationStatus.name,
     })));
     setRecordCount(data.recordCount);
-  }, [data?.results, data?.recordCount, isError, isLoading, analytics]);
+  }, [data?.directReviewsAvailable, data?.results, data?.recordCount, isError, isLoading, analytics]);
 
   useEffect(() => {
     if (data?.recordCount > 0 && pageNumber > 0 && data?.results?.length === 0) {
@@ -178,10 +154,10 @@ function ChplSearchView(props) {
     }
   }, [data?.recordCount, pageNumber, data?.results?.length]);
 
-  const downloadSearch = () => {
-    $analytics.eventTrack('Download Results', { category: analytics.category, label: listings.length });
-    csvExporter.generateCsv(listings);
-  };
+  useEffect(() => {
+    filterContext.dispatch('setFilterDisability', 'hasHadComplianceActivity', !directReviewsAvailable);
+    filterContext.dispatch('setFilterDisability', 'nonConformityOptions', !directReviewsAvailable);
+  }, [directReviewsAvailable]);
 
   const handleTableSort = (event, property, orderDirection) => {
     $analytics.eventTrack('Sort', { category: analytics.category, label: property });
@@ -205,6 +181,7 @@ function ChplSearchView(props) {
       <div className={classes.searchContainer} component={Paper}>
         <ChplFilterSearchTerm />
         <ChplFilterPanel />
+        <ChplFilterQuickFilters />
       </div>
       <div>
         <ChplFilterChips />
@@ -234,23 +211,10 @@ function ChplSearchView(props) {
               </div>
               { listings.length > 0
                 && (
-                  <ButtonGroup size="small" className={classes.wrap}>
-                    <Button
-                      color="secondary"
-                      variant="contained"
-                      fullWidth
-                      id="download-filtered-listings"
-                      onClick={downloadSearch}
-                    >
-                      Download
-                      {' '}
-                      { listings.length }
-                      {' '}
-                      Result
-                      { listings.length !== 1 ? 's' : '' }
-                      <GetAppIcon className={classes.iconSpacing} />
-                    </Button>
-                  </ButtonGroup>
+                  <ChplDownloadListings
+                    analytics={analytics}
+                    listings={listings}
+                  />
                 )}
             </div>
             { listings.length > 0